Finalize class
