+++ python/test.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import pypalmsens as ps
-from pypalmsens._instruments import capabilities2
 
 man = ps.connect()
 
@@ -9,8 +8,6 @@ comm = man._comm
 cap = man._comm.Capabilities
 
 
-c = capabilities2.Capabilities._init(comm=comm)
-
 if __name__ == '__main__':
+    c = man.capabilities
     breakpoint()
-    # c = man.capabilities
